requirements.txt
tensorflow==2.15.0
opencv-python-headless
numpy==1.26.4
pandas
scikit-learn
matplotlib
tqdm
Pillow
tensorflow-addons